apps/rag-service/requirements.txt
flask==3.1.0
gunicorn==23.0.0
openai==1.59.0
faiss-cpu==1.9.0
sentence-transformers==3.3.1
numpy==1.26.4
python-dotenv==1.0.1
